Switch dispatch ref to utilize the base of the PR
This change makes usage of the dispatch on label paradigm possible
at the expense of slowing development by forcing the running in the
base of the pull request rather than the merge ref. Ultimately this
may be more desirable as it enables the workflows run by self-hosted
runners to only execute trusted code from the base ref.